requirements.txt
python-valve
psutil
pyyaml
pysteamcmd
wheel
pywebview
flask
requests